+++ CHANGELOG.md
@@ -21,6 +21,7 @@ The format is simplified version of [Keep a Changelog](https://keepachangelog.co
 - Fix Recents page shows "No recent chapters" instead of a loading screen
 - Fix not fully loaded entries can't be selected on Library page
 - Fix certain Infinix devices being unable to use any "Open link in browser" actions, including tracker setup (@MajorTanya)
+- Fix source filter bottom sheet unable to be fully scrolled to the bottom
 
 ### Translation
 - Update translations from Weblate